浏览器初步

+++ puuman2/ViewAndController/Shop/RectHeadShowCell.h
@@ -20,7 +20,7 @@
     UILabel *wareNameLabel;
     UILabel *warePriceLabel;
     NSArray* wareArray;
-    int selectedIndex;
+    NSInteger selectedIndex;
   
 }
